pyproject.toml
[project]
name = "mcp-demos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-mcp-adapters>=0.1.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
